Manually fixed conflicts by merging GraphicsQualitySettings.cs into QualitySettings.cs
