--- Assets/Scripts/GraphicsQualitySettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60fbeba08c58841f6a765760c1e9d19c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
